Export compiler on its own entry
